Corrected alarm/kodi-c1-fb version
In Kodi's log, you can see that the version is actually 14.2-BETA2:
    NOTICE: Starting Kodi (14.2-BETA2 Git:Unknown). Platform: Linux ARM
    32-bit
Thus, the version stated in this PKGBUILD is incorrect.
I changed the pkgver to "14.2" and the pkgrel to "0.beta2.1". If the
final 14.2 version is reached, we can safely set pkgrel to 1. As long
as we stay with 14.2-BETA2, we can just increment the last number,
i.e. "0.beta2.2".